src/main/python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hrv-bridge"
version = "1.0.0"
description = "MQTT to PWM/DAC bridge for HRV power output control"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Robert Fiser"}
]
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Home Automation",
]
dependencies = [
    "paho-mqtt>=1.6.0,<3.0.0",
    "spidev>=3.0",
    "lgpio>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
hrv-bridge = "dac_bridge:main"

[tool.setuptools]
packages = ["dac_bridge"]
